refactor(meta): split iceberg compaction manager module
