+++ src/meta/src/manager/iceberg_compaction/manual.rs
@@ -1,0 +1,154 @@
+// Copyright 2025 RisingWave Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use anyhow::anyhow;
+use iceberg::spec::Operation;
+use risingwave_connector::sink::catalog::SinkId;
+use risingwave_connector::sink::iceberg::should_enable_iceberg_cow;
+use risingwave_pb::iceberg_compaction::IcebergCompactionTask;
+use risingwave_pb::iceberg_compaction::iceberg_compaction_task::TaskType;
+
+use super::*;
+
+impl IcebergCompactionManager {
+    pub async fn trigger_manual_compaction(&self, sink_id: SinkId) -> MetaResult<u64> {
+        use risingwave_pb::iceberg_compaction::subscribe_iceberg_compaction_event_response::Event as IcebergResponseEvent;
+
+        let iceberg_config = self.load_iceberg_config(sink_id).await?;
+        let initial_table = iceberg_config.load_table().await?;
+        let initial_snapshot_id = initial_table
+            .metadata()
+            .current_snapshot()
+            .map(|s| s.snapshot_id())
+            .unwrap_or(0);
+        let initial_timestamp = chrono::Utc::now().timestamp_millis();
+
+        let compactor = self
+            .iceberg_compactor_manager
+            .next_compactor()
+            .ok_or_else(|| anyhow!("No iceberg compactor available"))?;
+
+        let task_id = self
+            .env
+            .hummock_seq
+            .next_interval("compaction_task", 1)
+            .await?;
+
+        let sink_param = self.get_sink_param(sink_id).await?;
+
+        compactor.send_event(IcebergResponseEvent::CompactTask(IcebergCompactionTask {
+            task_id,
+            sink_id: sink_id.as_raw_id(),
+            props: sink_param.properties,
+            task_type: TaskType::Full as i32,
+        }))?;
+
+        tracing::info!(
+            "Manual compaction triggered for sink {} with task ID {}, waiting for completion...",
+            sink_id,
+            task_id
+        );
+
+        self.wait_for_compaction_completion(
+            &sink_id,
+            iceberg_config,
+            initial_snapshot_id,
+            initial_timestamp,
+            task_id,
+        )
+        .await?;
+
+        Ok(task_id)
+    }
+
+    async fn wait_for_compaction_completion(
+        &self,
+        sink_id: &SinkId,
+        iceberg_config: IcebergConfig,
+        initial_snapshot_id: i64,
+        initial_timestamp: i64,
+        task_id: u64,
+    ) -> MetaResult<()> {
+        const INITIAL_POLL_INTERVAL_SECS: u64 = 2;
+        const MAX_POLL_INTERVAL_SECS: u64 = 60;
+        const MAX_WAIT_TIME_SECS: u64 = 1800;
+        const BACKOFF_MULTIPLIER: f64 = 1.5;
+
+        let mut elapsed_time = 0;
+        let mut current_interval_secs = INITIAL_POLL_INTERVAL_SECS;
+
+        let cow =
+            should_enable_iceberg_cow(iceberg_config.r#type.as_str(), iceberg_config.write_mode);
+
+        while elapsed_time < MAX_WAIT_TIME_SECS {
+            let poll_interval = std::time::Duration::from_secs(current_interval_secs);
+            tokio::time::sleep(poll_interval).await;
+            elapsed_time += current_interval_secs;
+
+            tracing::info!(
+                "Checking iceberg compaction completion for sink {} task_id={}, elapsed={}s, interval={}s",
+                sink_id,
+                task_id,
+                elapsed_time,
+                current_interval_secs
+            );
+
+            let current_table = iceberg_config.load_table().await?;
+
+            let metadata = current_table.metadata();
+            let new_snapshots: Vec<_> = metadata
+                .snapshots()
+                .filter(|snapshot| {
+                    let snapshot_timestamp = snapshot.timestamp_ms();
+                    let snapshot_id = snapshot.snapshot_id();
+                    snapshot_timestamp > initial_timestamp && snapshot_id != initial_snapshot_id
+                })
+                .collect();
+
+            for snapshot in new_snapshots {
+                let summary = snapshot.summary();
+                if cow {
+                    if matches!(summary.operation, Operation::Overwrite) {
+                        tracing::info!(
+                            "Iceberg compaction completed for sink {} task_id={} with Overwrite operation",
+                            sink_id,
+                            task_id
+                        );
+                        return Ok(());
+                    }
+                } else if matches!(summary.operation, Operation::Replace) {
+                    tracing::info!(
+                        "Iceberg compaction completed for sink {} task_id={} with Replace operation",
+                        sink_id,
+                        task_id
+                    );
+                    return Ok(());
+                }
+            }
+
+            current_interval_secs = std::cmp::min(
+                MAX_POLL_INTERVAL_SECS,
+                ((current_interval_secs as f64) * BACKOFF_MULTIPLIER) as u64,
+            );
+        }
+
+        Err(anyhow!(
+            "Iceberg compaction did not complete within {} seconds for sink {} (task_id={})",
+            MAX_WAIT_TIME_SECS,
+            sink_id,
+            task_id
+        )
+        .into())
+    }
+}
