+++ src/meta/src/manager/iceberg_compaction/stream.rs
@@ -1,0 +1,90 @@
+// Copyright 2025 RisingWave Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::sync::Arc;
+
+use risingwave_common::id::WorkerId;
+use risingwave_connector::connector_common::IcebergSinkCompactionUpdate;
+use risingwave_pb::iceberg_compaction::SubscribeIcebergCompactionEventRequest;
+use tokio::sync::mpsc::UnboundedReceiver;
+use tokio::sync::oneshot::Sender;
+use tokio::task::JoinHandle;
+use tonic::Streaming;
+
+use super::*;
+use crate::hummock::{
+    IcebergCompactionEventDispatcher, IcebergCompactionEventHandler, IcebergCompactionEventLoop,
+};
+
+impl IcebergCompactionManager {
+    pub fn compaction_stat_loop(
+        manager: Arc<Self>,
+        mut rx: UnboundedReceiver<IcebergSinkCompactionUpdate>,
+    ) -> (JoinHandle<()>, Sender<()>) {
+        let (shutdown_tx, mut shutdown_rx) = tokio::sync::oneshot::channel();
+        let join_handle = tokio::spawn(async move {
+            loop {
+                tokio::select! {
+                    Some(stat) = rx.recv() => {
+                        manager.update_iceberg_commit_info(stat).await;
+                    },
+                    _ = &mut shutdown_rx => {
+                        tracing::info!("Iceberg compaction manager is stopped");
+                        return;
+                    }
+                }
+            }
+        });
+
+        (join_handle, shutdown_tx)
+    }
+
+    pub fn add_compactor_stream(
+        &self,
+        context_id: WorkerId,
+        req_stream: Streaming<SubscribeIcebergCompactionEventRequest>,
+    ) {
+        if self
+            .compactor_streams_change_tx
+            .send((context_id, req_stream))
+            .is_err()
+        {
+            tracing::warn!(context_id = %context_id, "Failed to enqueue iceberg compactor stream");
+        }
+    }
+
+    pub fn iceberg_compaction_event_loop(
+        iceberg_compaction_manager: Arc<Self>,
+        compactor_streams_change_rx: CompactorChangeRx,
+    ) -> Vec<(JoinHandle<()>, Sender<()>)> {
+        let mut join_handle_vec = Vec::default();
+
+        let iceberg_compaction_event_handler =
+            IcebergCompactionEventHandler::new(iceberg_compaction_manager.clone());
+
+        let iceberg_compaction_event_dispatcher =
+            IcebergCompactionEventDispatcher::new(iceberg_compaction_event_handler);
+
+        let event_loop = IcebergCompactionEventLoop::new(
+            iceberg_compaction_event_dispatcher,
+            iceberg_compaction_manager.metrics.clone(),
+            compactor_streams_change_rx,
+        );
+
+        let (event_loop_join_handle, event_loop_shutdown_tx) = event_loop.run();
+        join_handle_vec.push((event_loop_join_handle, event_loop_shutdown_tx));
+
+        join_handle_vec
+    }
+}
